chore: pull num-complex with no default features
num-complex's default features include std, which is wrong for a
workspace that publishes no-std crates. Set default-features = false
at the workspace declaration and pull in libm explicitly so float
arithmetic still works in no-std.

The Falcon DSA code called std-only inherent f64 methods (cos, sin,
round, sqrt, floor). Switch those call sites to num::Float trait
methods, which resolve to libm-backed implementations in no-std and
to the inherent methods in std. The num crate already had the libm
feature, so the trait was available without further dependency changes.

+++ crates/crypto/src/dsa/falcon512_poseidon2/math/fft.rs
@@ -4,7 +4,7 @@ use core::{
     ops::{Add, AddAssign, Mul, MulAssign, Neg, Sub, SubAssign},
 };
 
-use num::{One, Zero};
+use num::{Float, One, Zero};
 use num_complex::Complex64;
 
 use super::{Inverse, field::FalconFelt, polynomial::Polynomial};
@@ -218,7 +218,7 @@ where
 impl CyclotomicFourier for Complex64 {
     fn primitive_root_of_unity(n: usize) -> Self {
         let angle = 2. * PI / (n as f64);
-        Complex64::new(f64::cos(angle), f64::sin(angle))
+        Complex64::new(Float::cos(angle), Float::sin(angle))
     }
 
     /// Custom implementation of CyclotomicFourier::bitreversed_powers for
@@ -228,7 +228,7 @@ impl CyclotomicFourier for Complex64 {
         let half_circle = PI;
         for (i, a) in array.iter_mut().enumerate() {
             let angle = (i as f64) * half_circle / (n as f64);
-            *a = Self::new(f64::cos(angle), f64::sin(angle));
+            *a = Self::new(Float::cos(angle), Float::sin(angle));
         }
         Self::bitreverse_array(&mut array);
         array
@@ -241,7 +241,7 @@ impl CyclotomicFourier for Complex64 {
         let half_circle = PI;
         for (i, a) in array.iter_mut().enumerate() {
             let angle = (i as f64) * half_circle / (n as f64);
-            *a = Self::new(f64::cos(angle), -f64::sin(angle));
+            *a = Self::new(Float::cos(angle), -Float::sin(angle));
         }
         Self::bitreverse_array(&mut array);
         array
